Task/fix import indexing (#159)
* Issue #157: Add test case to show bug

* builds

* add gimli to cargo.toml

* Issue #157: Add more tests and remove some unnecessary code

* Issue #157: Use generics for reindexing logic

* Issue #157: Recalculate global IDs

* Issue #157: work on the API

* Issue #157: Add lots of testing

* Issue #157: Add lots of testing

---------

+++ src/subiterator/component_subiterator.rs
@@ -34,7 +34,7 @@ impl ComponentSubIterator {
             num_mods,
             metadata: metadata.clone(),
             mod_iterator: ModuleSubIterator::new(
-                metadata.get(&0).unwrap().keys().len(),
+                metadata.get(&0).unwrap().keys().len() as u32,
                 (*metadata.get(&curr_mod).unwrap()).clone(),
                 match skip_funcs.contains_key(&curr_mod) {
                     true => skip_funcs.get(&curr_mod).unwrap().clone(),
@@ -56,7 +56,7 @@ impl ComponentSubIterator {
     fn next_module(&mut self) -> bool {
         self.curr_mod += 1;
         if self.curr_mod < self.num_mods as u32 {
-            let num_funcs = self.metadata.get(&self.curr_mod).unwrap().keys().len();
+            let num_funcs = self.metadata.get(&self.curr_mod).unwrap().keys().len() as u32;
             let met = self.metadata.get(&self.curr_mod).unwrap().clone();
             // If we're defining a new module, we have to reset function
             self.mod_iterator = ModuleSubIterator::new(
